Rename 'element-stuff' to something more appropriate
This module was exporting a number of DOM helpers used across the
modules we currently have.

This renames it and moves it out of deprecated as it is quite heavily
used in the current codebase so can't really be called deprecated
until we move to an alternative way of working with the DOM.
