+++ assets/javascripts/_deprecated/lookup-address.js
@@ -1,6 +1,6 @@
 /* eslint no-new: 0 */
 const axios = require('axios')
-const { closest, hide } = require('./lib/element-stuff')
+const { closest, hide } = require('../lib/helpers')
 
 class LookupAddress {
   constructor (elementSelector) {
