+++ assets/javascripts/lib/helpers.js
