+++ assets/javascripts/modules/archive-form.js
@@ -1,4 +1,4 @@
-const { hide, show, addClass, insertAfter } = require('../_deprecated/lib/element-stuff')
+const { hide, show, addClass, insertAfter } = require('../lib/helpers')
 
 /**
  * ArchiveForm
