+++ assets/javascripts/modules/sortable-table.js
@@ -1,7 +1,7 @@
 const uniqueId = require('lodash/uniqueId')
 
 const { parseDateString } = require('../../../common/date')
-const { addClass, removeClass } = require('../_deprecated/lib/element-stuff')
+const { addClass, removeClass } = require('../lib/helpers')
 
 function isDate (text) {
   return parseDateString(text) !== null
